+++ CHANGELOG.md
@@ -6,14 +6,25 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
-## [3.1] - 2022-04-18
+## [3.2] -2023-05-22
+
+### Added
+
+* NFS client and server settings
+
+### Fixed
+
+* Check only running instances of MSSQL and make the registry settings check more error proof
+* Changed ``Write-Information`` cmdlet to ``Write-Output`` to be backward compatible
+
+## [3.1] - 2023-04-18
 
 ### Added
 
 * Registry value check for Kernel DMA Protection
 * Check for MSSQL connection settings
 
-## [3.0] - 2022-03-09
+## [3.0] - 2023-03-09
 
 ### Added